Add files via upload
Use the robust functionality of Pandas to extract the required data from the CSV file, calculate the exponentiation of the data and tabulate it. This method can provide us with the data to be analyzed more intuitively. Also, use the visualization capabilities of Matplotlib to transform the data into pivot tables or charts. Graphics are a way to communicate across language barriers, and visualizing data is necessary to simplify the analysis process.